GHOSTCORE/requirements.txt
reportlab
python-docx
pyyaml
Pillow
qrcode
